// src/types.ts
export type UserRole = 'OWNER' | 'BRANCH_MANAGER' | 'CASHIER' | 'SALES_STAFF';

export type PetAvatarType = 'dog' | 'cat' | 'bird' | 'fish' | 'rabbit' | 'hamster';

export interface User {
  id: string;
  username: string;
  name: string;
  email?: string;
  phone?: string;
  role: UserRole;
  branchId?: string; // If null/empty -> Owner/Super Admin
  avatarType: PetAvatarType;
  designation?: string;
  status?: 'ACTIVE' | 'INACTIVE';
}

export interface Branch {
  id: string;
  code: string;
  name: string;
  address: string;
  city: string;
  phone: string;
  email: string;
  managerName: string;
  status: 'ACTIVE' | 'INACTIVE';
  openingDate: string;
  taxRate: number;
  gstin?: string;
  accessPassword?: string;
}